health_recipe/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-cors-headers==3.2.1
django-simpleui==4.0.2
django-summernote==0.8.11.6
djangorestframework==3.11.0
idna==2.9
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.4
pandas==0.25.3
Pillow==7.1.2
prettytable==0.7.2
pyecharts==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
xlrd==1.2.0
